requirements.txt
langchain
openai
faiss-cpu
tqdm
numpy
python-dotenv
qdrant-client
datasets
ragas
scikit-learn
ipykernel
jupyter
pandas
langchain_community
langchain_openai
pathlib
langchain-anthropic
langchain-qdrant
langchain-core
langchain-openai